+++ src/application/overwrites/overwriteRequestHeaders.test.ts
@@ -120,4 +120,18 @@ describe('overwriteRequestHeaders', () => {
     const result = overwriteRequestHeaders(overwriteValues, pmOperation)
     expect(result.item.request.getHeaders()).toMatchSnapshot()
   })
+
+  it('should overwrite the request headers number variable with string value', async () => {
+    const overwriteValues = [
+      {
+        key: 'x-apideck-app-id',
+        value: -1
+      }
+    ]
+    const pmOperation = await getPostmanMappedOperation()
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    const result = overwriteRequestHeaders(overwriteValues, pmOperation)
+    expect(result.item.request.getHeaders()).toMatchSnapshot()
+  })
 })