+++ src/application/overwrites/overwriteRequestPathVariables.test.ts
@@ -40,4 +40,19 @@ describe('overwriteRequestPathVariables', () => {
 
     expect(result.item.request.url.variables).toEqual(pmOperation.item.request.url.variables)
   })
+
+  it('should overwrite the request path number variable with string value', async () => {
+    const overwriteValues = [
+      {
+        key: 'id',
+        value: -1
+      }
+    ]
+
+    const pmOperation = await getPostmanMappedOperation()
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    const result = overwriteRequestPathVariables(overwriteValues, pmOperation)
+    expect(result.item.request.url.variables).toMatchSnapshot()
+  })
 })