+++ src/application/overwrites/overwriteRequestQueryParams.test.ts
@@ -119,4 +119,20 @@ describe('overwriteRequestQueryParams', () => {
     const result = overwriteRequestQueryParams(overwriteValues, pmOperation)
     expect(result.item.request.url.query).toMatchSnapshot()
   })
+
+  it('should overwrite the query param number variable with string value', async () => {
+    const overwriteValues = [
+      {
+        key: 'raw',
+        value: -1,
+        overwrite: true
+      }
+    ]
+
+    const pmOperation = await getPostmanMappedOperation()
+    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+    // @ts-ignore
+    const result = overwriteRequestQueryParams(overwriteValues, pmOperation)
+    expect(result.item.request.url.query).toMatchSnapshot()
+  })
 })